ADD :: logout query feature로 뺌
